Add ability to exclude resources from Buck resource map
Summary: Resources marked as `exclude-from-buck-resource-map="true"` will not be used to build the R.txt map for checking Buck dependencies. This has the effect of breaking the build if the original resource is removed, but a generated resource is still being referenced by code. You don't need to regenerate the generated resources to find the build break.

Test Plan: Annotate generated resources with top level attribute `exclude-from-buck-resource-map="true"`, see the build breaks when I remove the source English string now.