Split up preprocess related functionality of CxxPreprocessAndCompile rule
Summary:
This create a delegate class that handles Preprocessor related logic for
CxxPreprocessAndCompile rule. The eventual goal of that starts with this is
breaking up CxxPreprocessAndCompile rule into smaller rules that are more
maintainable and testable.

This diff essentially rips out the preprocessor specific logic, and replace
method calls to this logic with calls to the delegate. The delegate contributes
to the rule key as the fields used to do. In this sense the operation should be
a no-op.

This also fixes a broken case (the tests were broken too) in CompilationDB
generation.

Test Plan: Tests pass.