[xcode] Don't escape linker flags in OTHER_LDFLAGS
Summary:
On my machine, escaping the linker flags results in the following warning from
clang when building a project with xcodebuild that uses a halide_library target,
and which was generated using 'buck project --combine-test-bundles':

  clang: warning: argument unused during compilation: '-Xlinker libfoo.a -Xlinker -lc++ -Xlinker -lstdc++ -Xlinker libfoo.a -Xlinker -lc++ -Xlinker -lstdc++'

(Where I've substituted 'libfoo.a' for the path to the library generated by my
test halide_library target.) The resulting build fails because the symbols
defined in libfoo.a can't be found. Removing the code for shell escaping the
linker flags fixes these issues.

Test Plan:
Ran 'buck project --combine-test-target' for my test project which includes a
halide_library target. Built the project and ran the unit tests for it, both
in Xcode and on the command-line using xcodebuild.