Easy: Ignore SIGPIPE when Python nailgun client sends heartbeat
Summary: When shutting down the Python nailgun client, we might
need to send a heartbeat during the shutdown process.

If the server closes its socket while this is happening, we
would get SIGPIPE and crash the Python nailgun client.

The C client handles this by ignoring SIGPIPE on heartbeats,
so we will do the same in the Python client.

Test Plan: Add a `throw new RuntimeException()` to Buck. Run
`buck build //path/to:target`. Previously, Python client
would crash with SIGPIPE. Now, it cleanly shuts down.

Reviewed By: Coneko

fb-gh-sync-id: a00612b