[halide] Update halide-compile.st to support more platforms
Summary: Check for a supported platform in halide-compile.st, using the PLATFORM_NAME
environment variable. For unsupported platforms, we build an 'empty' library for
now, so that we still have something to link at link-time (ld returns an error
if the Halide library is missing).

Test Plan: Ran 'buck project' for a target which transitively uses a halide_library; built
the resulting project.

Reviewed By: vladb

fb-gh-sync-id: 218c21e