Add a new `Manifest` data structure for manifest-based caching
Summary: This adds a new `Manifest` class which handles most of the functionality
needed for manifest-based caching.  Manifest-based caching attempts to
serve cache hits using the minimum set of inputs without knowing the set
of inputs beforehand.  It accomplishes this by storing a mapping of various
states of the inputs to another cache key that can be used to lookup the
output if the clients on-disk state matches the inputs.  Therefore, at a
high-level, the manifest is basically a `Map<RuleKey, Map<Path, HashCode>`.
When a lookup is performed, the manifest entries are searched and the stored
path hashes are compared to what actually exists on-disk.  If a hit is found,
the associated key is used to look back in the cache for the real output.

This implementation is designed so that it supports the following operations
efficiently:
- Serialize the manifest to an on-disk format for shipping to the cache.
- Deserialize the manifest from an on-disk format for loading form the cache.
- Lookup a match using the on-disk state represented by a `FileHashCache`.
- Append a new entry to the end of the manifest.

Test Plan: added unittests

Reviewed By: Coneko

fb-gh-sync-id: 036b743