CachingBuildEngine: enable manifest-based caching
Summary: This implements manifest-based caching for rules which support dep-files
in the build engine.  With this change, on dep-file miss, the build engine
downloads the associated manifest and performs a lookup using any hits
found within.  After a rule is built locally, the manifest is updated with
the resulting entry and the artifact is also indexed into the cache using
the dep-file rule key.

Test Plan: added unittests

Reviewed By: Coneko

fb-gh-sync-id: 66096bc