Ensure unhandled exceptions on background threads crash Buck
Summary: The default Java behavior of unhandled exceptions is
to print a stack trace to `System.err` then terminate the thread.
However, the rest of the JVM continues running.

If the terminated thread held a resource which another thread
needs, Buck will deadlock.

Especially in buckd Nailgun environments, it's not guaranteed
that printing to `System.err` will ever make its way to the
client (it has to write to the Nailgun socket, and who knows
what state that is in after an exception went unhandled).

This diff introduces a global unhandled exception handler
in `Main.java` which logs a stack trace to the console
and the log file, shuts down Nailgun if it's running,
then exits the JVM with error code 1.

Test Plan: Added to Main.java:

      Thread t = new Thread() {
          Override
          public void run() {
            throw new RuntimeException("oops");
          }
        };
      t.start();

Ran `buck version` and `NO_BUCKD=1 buck version`. Previously,
both of them just printed a stack trace and continued running.

Now, the JVM exits with error code 1 in both cases.

Reviewed By: nayrmi

fb-gh-sync-id: fdea8a5