Cache Cells' hash codes
Summary:
Since `Cells` are used as keys of various maps, they are fairly often
hashed. Since all of their fields are final, their hash can be cached
and doing so significantly increases parsing performance.

Test Plan:
`buck test`. Run `buck targets --show-output ${TARGET}` with and
without this change and see that the performance is about 40% better
with it.

Reviewed By: Coneko

fb-gh-sync-id: d7f9be4