Ant build: Compile AOSP classes into their own directory with ecj
Summary:
Previously, the Ant build compiled the Android open-source
classes into `build/classes` with the rest of Buck.

This makes `ant forbidden-checks` less useful, since it doesn't know
which classes are Buck's and which are AOSP's (and AOSP's have a lot
of naughty API calls).

This diff consolidates the AOSP compile to just compile with ecj
(which we were already doing for the Immutables build) and points
`buck_repo.py` at the location of the classes.

Test Plan:
Clean and incremental ant builds.
  buck test //test/com/facebook/buck/android/aapt:
  ant java-test -Dtest.class=MiniAaptTest
  ant forbidden-checks

Previously, lots of forbidden API checks hit on `com.android.*`.
Now, they don't.

Reviewed By: sdwilsh

fb-gh-sync-id: 5c537ec