Correctly post rule key computation events
Summary:
We were passing the rule key to the `BuildRuleEvent.Started` event,
which meant that we needed to compute the rule key before the event
was posted, so the rule key computation occurred outside of the event
pair. Fix this by removing information about the rule key from
`BuildRuleEvent.Started`.

Test Plan:
`buck test`.
Run a build and confirm that the Chrome traces now cover the whole
interval for computing rule keys and that the computed rule keys are
still displayed.

Reviewed By: marcinkosiba

fb-gh-sync-id: 1fea51a