[react-native-packager] Add sourcemap option for unbundles, too
Summary:
As ct-scan expects the react native packager to write a source map in call cases, we pass the corresponding even when building unbundles.

Unbundles in dev mode don’t need a source map, but we will rely on a source map for release builds for live crash reports.

Test Plan:
- adapt and run integration tests
- build rnfeed/android and verify that the packager writes that map

Reviewed By: martinbigio

fb-gh-sync-id: c720c0d