Use a `CommandThreadManager` in `InstallCommand`
Summary:
A future change will require an `Executor` in `InstallCommand`.  This
change prepares us for this by passing one into the `BuildCommand`
instead of relying on it to create one for us.

This also names our threads more inline with how we call them
elsewhere: by the command name.

Test Plan: CI

Reviewed By: andrewjcg

fb-gh-sync-id: c1ca725