Don't check that we've passed at least one target twice
Summary:
In the install command, we'd check if we were given any targets, and
then call into the build command, which would do the same thing (but
with a better error message).  This change unifies the two into a
helper method in the build command in order to share this check.  A
future change will make us only run it once.

Test Plan: CI

Reviewed By: shs96c

fb-gh-sync-id: 69f4428