Fix a bug with publishing artifacts to maven
Maven artifacts can be thought of as being public-facing
APIs, but you don't always want to structure your build
files that way. To resolve that problem, it'd be possible
to create a single `java_library` that simply exports
everything that's needed, and ensure that dependent libraries
use that.

Sadly, if this is done prior to this fix, then dependent
libraries with maven_coords wouldn't be published properly.

As a worked example, consider an "api" jar, and an "impl"
jar. Obviously "impl" depends on "api". But "api" is defined
as:

java_library(name = 'api',
  exported_deps = [':a', ':b'],
  maven_coords = 'com.example:api:1.0',
)

java_library(name = 'impl',
  srcs = glob(['*.java']),
  deps = [':api'],
  maven_coords = 'com.example:impl:1.0',
)

Prior to this diff, the contents of ':a' and ':b' would be
in the 'impl' artifact. D'oh!